Block Texturing
Nope didn't make any textures. Just improved my block texturing system
so that now it rotates the block the right way when you place it, and it
renders right in your inventory. I think next update will be upgrading
the gui system (because it sucks right now). Then i'll look into making
the mod structure (and so actualy making it playable! Might be close to
the first release!)